chore(ui) Extract more types from types/index
Pull out some of the organization type and integration/plugin/sentry app
types. I've opted to have a single larger integrations module over small
separate modules for sentryapps/integrations/plugins as I didn't want to
add more imports and have to find homes for resources that are shared
across the three integration types.
