+++ static/app/types/auth.tsx
@@ -98,3 +98,23 @@ export type UserEnrolledAuthenticator = {
   id: EnrolledAuthenticator['authId'];
   name: EnrolledAuthenticator['name'];
 };
+
+/**
+ * XXX(ts): This actually all comes from getsentry. We should definitely
+ * refactor this into a more proper 'hook' mechanism in the future
+ */
+export type AuthConfig = {
+  canRegister: boolean;
+  serverHostname: string;
+  hasNewsletter: boolean;
+  githubLoginLink: string;
+  vstsLoginLink: string;
+  googleLoginLink: string;
+};
+
+export type AuthProvider = {
+  key: string;
+  name: string;
+  requiredFeature: string;
+  disables2FA: boolean;
+};
