+++ static/app/types/user.tsx
@@ -1,4 +1,6 @@
+import {Authenticator, EnrolledAuthenticator} from './auth';
 import {Avatar} from './core';
+import {UserExperiments} from './experiments';
 
 /**
  * Avatars are a more primitive version of User.
@@ -17,4 +19,55 @@ export type AvatarUser = {
     avatarType: Avatar['avatarType'];
   };
   lastSeen?: string;
+};
+
+/**
+ * This is an authenticator that a user is enrolled in
+ */
+type UserEnrolledAuthenticator = {
+  dateUsed: EnrolledAuthenticator['lastUsedAt'];
+  dateCreated: EnrolledAuthenticator['createdAt'];
+  type: Authenticator['id'];
+  id: EnrolledAuthenticator['authId'];
+  name: EnrolledAuthenticator['name'];
+};
+
+export type User = Omit<AvatarUser, 'options'> & {
+  lastLogin: string;
+  isSuperuser: boolean;
+  isAuthenticated: boolean;
+  emails: {
+    is_verified: boolean;
+    id: string;
+    email: string;
+  }[];
+  isManaged: boolean;
+  lastActive: string;
+  isStaff: boolean;
+  identities: any[];
+  isActive: boolean;
+  has2fa: boolean;
+  canReset2fa: boolean;
+  authenticators: UserEnrolledAuthenticator[];
+  dateJoined: string;
+  options: {
+    theme: 'system' | 'light' | 'dark';
+    timezone: string;
+    stacktraceOrder: number;
+    language: string;
+    clock24Hours: boolean;
+    avatarType: Avatar['avatarType'];
+  };
+  flags: {newsletter_consent_prompt: boolean};
+  hasPasswordAuth: boolean;
+  permissions: Set<string>;
+  experiments: Partial<UserExperiments>;
+};
+
+// XXX(epurkhiser): we should understand how this is diff from User['emails]
+// above
+export type UserEmail = {
+  email: string;
+  isPrimary: boolean;
+  isVerified: boolean;
 };
